✨ 지원자 현황 접근 상태 분기 추가
