+++ apps/web/src/apis/Scores/getLanguageTestList.ts
@@ -5,10 +5,11 @@ import { ScoresQueryKeys, scoresApi } from "./api";
 /**
  * @description 내 어학 점수 조회 훅
  */
-const useGetMyLanguageTestScore = () => {
+const useGetMyLanguageTestScore = ({ enabled = true }: { enabled?: boolean } = {}) => {
   return useQuery({
     queryKey: [ScoresQueryKeys.myLanguageTestScore],
     queryFn: scoresApi.getMyLanguageTestScore,
+    enabled,
     staleTime: Infinity,
     select: (data) => data.data.languageTestScoreStatusResponseList,
   });
