+++ apps/web/src/app/university/application/ScorePageContent.tsx
@@ -1,420 +1,46 @@
 "use client";
 
-import clsx from "clsx";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { type ReactNode, useEffect, useMemo, useState } from "react";
-import { useGetApplicationsList } from "@/apis/applications";
-import ConfirmCancelModal from "@/components/modal/ConfirmCancelModal";
-import { DEFAULT_MAX_CHOICE_COUNT, getHomeUniversityById, REGIONS_KO } from "@/constants/university";
+import { useGetMyGpaScore, useGetMyLanguageTestScore } from "@/apis/Scores";
+import CloudSpinnerPage from "@/components/ui/CloudSpinnerPage";
 import useAuthStore from "@/lib/zustand/useAuthStore";
-import { IconExpandMoreFilled } from "@/public/svgs/community";
-import type { Applicant, ScoreSheet as ScoreSheetType } from "@/types/application";
-import type { RegionKo } from "@/types/university";
-import { getApplicationDetailHref, MobileScoreSheet, ScoreSheetLogo } from "./ScoreSheet";
-
-type ApplicantScope = "all" | "withApplicants";
-type ScoreSort = "applicants" | "gpa";
-
-type AppliedUniversity = {
-  preference: number;
-  scoreSheet: ScoreSheetType;
-};
-
-type ScorePageViewProps = {
-  appliedUniversities: AppliedUniversity[];
-  displayedScoreSheets: ScoreSheetType[];
-  totalUniversityCount: number;
-  applicantUniversityCount: number;
-  participantCount: number;
-  scope: ApplicantScope;
-  regionFilter: RegionKo | "";
-  sortMode: ScoreSort;
-  onScopeChange: (scope: ApplicantScope) => void;
-  onRegionChange: (region: RegionKo | "") => void;
-  onSortModeChange: (sortMode: ScoreSort) => void;
-  onChangeUniversities: () => void;
-};
+import ApprovedApplicationStatusPage from "./_pages/ApprovedApplicationStatusPage";
+import { resolveApplicationStatusAccess } from "./_pages/applicationStatusAccess";
+import GuestApplicationStatusPage from "./_pages/GuestApplicationStatusPage";
+import ScorePendingApplicationStatusPage from "./_pages/ScorePendingApplicationStatusPage";
+import SignedInApplicationStatusPage from "./_pages/SignedInApplicationStatusPage";
 
 const ScorePageContent = () => {
-  const router = useRouter();
+  const isAuthInitialized = useAuthStore((state) => state.isInitialized);
+  const isAuthenticated = useAuthStore((state) => state.isAuthenticated);
   const homeUniversityId = useAuthStore((state) => state.homeUniversityId);
-  const maxChoiceCount = getHomeUniversityById(homeUniversityId)?.maxChoiceCount ?? DEFAULT_MAX_CHOICE_COUNT;
+  const shouldFetchScoreStatus = isAuthInitialized && isAuthenticated && homeUniversityId !== null;
 
-  const [scope, setScope] = useState<ApplicantScope>("withApplicants");
-  const [regionFilter, setRegionFilter] = useState<RegionKo | "">("");
-  const [sortMode, setSortMode] = useState<ScoreSort>("applicants");
-  const [showNeedApply, _setShowNeedApply] = useState(false);
-
-  const emptyChoices = useMemo(
-    () => Array.from({ length: maxChoiceCount }, () => [] as ScoreSheetType[]),
-    [maxChoiceCount],
-  );
-  const { data: scoreResponseData, isError, isLoading } = useGetApplicationsList();
-  const scoreChoices = scoreResponseData?.choices ?? emptyChoices;
-
-  const allScoreSheets = useMemo(
-    () => sortScoreSheets(uniqueScoreSheets(scoreChoices.flat()), sortMode),
-    [scoreChoices, sortMode],
-  );
-  const appliedUniversities = useMemo(
-    () => getAppliedUniversities(scoreChoices, maxChoiceCount),
-    [scoreChoices, maxChoiceCount],
-  );
-  const totalUniversityCount = allScoreSheets.length;
-  const applicantUniversityCount = allScoreSheets.filter((scoreSheet) => scoreSheet.applicants.length > 0).length;
-  const participantCount = useMemo(() => getParticipantCount(allScoreSheets), [allScoreSheets]);
-
-  const displayedScoreSheets = useMemo(() => {
-    let result =
-      scope === "withApplicants"
-        ? allScoreSheets.filter((scoreSheet) => scoreSheet.applicants.length > 0)
-        : allScoreSheets;
-
-    if (regionFilter) {
-      result = result.filter((scoreSheet) => scoreSheet.region === regionFilter);
-    }
-
-    return sortScoreSheets(result, sortMode);
-  }, [allScoreSheets, regionFilter, scope, sortMode]);
-
-  useEffect(() => {
-    if (isLoading) return;
-    if (isError) {
-      router.replace("/university/application/apply");
-    }
-  }, [isError, isLoading, router]);
-
-  const viewProps = {
-    appliedUniversities,
-    displayedScoreSheets,
-    totalUniversityCount,
-    applicantUniversityCount,
-    participantCount,
-    scope,
-    regionFilter,
-    sortMode,
-    onScopeChange: setScope,
-    onRegionChange: setRegionFilter,
-    onSortModeChange: setSortMode,
-    onChangeUniversities: () => router.push("/university/application/apply"),
-  };
-
-  return (
-    <>
-      <ApplicationScoreView {...viewProps} />
-      <ConfirmCancelModal
-        title="학교 지원이 필요합니다"
-        isOpen={showNeedApply}
-        handleCancel={() => router.push("/")}
-        handleConfirm={() => router.push("/university/application/apply")}
-        content={"점수 공유현황을 확인하려면 지원절차를\n진행해주세요."}
-        cancelText="확인"
-        approveText="학교 지원하기"
-      />
-    </>
-  );
-};
-
-const ApplicationScoreView = ({
-  appliedUniversities,
-  displayedScoreSheets,
-  totalUniversityCount,
-  applicantUniversityCount,
-  participantCount,
-  scope,
-  regionFilter,
-  sortMode,
-  onScopeChange,
-  onRegionChange,
-  onSortModeChange,
-  onChangeUniversities,
-}: ScorePageViewProps) => {
-  return (
-    <main className="mx-auto w-full max-w-app px-5 pb-6 md:pt-14">
-      <AppliedUniversitySection appliedUniversities={appliedUniversities} onChangeUniversities={onChangeUniversities} />
-      <div className="mt-5 h-1 bg-k-50" />
-      <ParticipantBanner participantCount={participantCount} />
-      <ApplicationScopeTabs
-        scope={scope}
-        totalUniversityCount={totalUniversityCount}
-        applicantUniversityCount={applicantUniversityCount}
-        onScopeChange={onScopeChange}
-      />
-      <ApplicationFilterChips
-        regionFilter={regionFilter}
-        sortMode={sortMode}
-        onRegionChange={onRegionChange}
-        onSortModeChange={onSortModeChange}
-      />
-      <ScoreSheetList scoreSheets={displayedScoreSheets} />
-    </main>
-  );
-};
-
-const AppliedUniversitySection = ({
-  appliedUniversities,
-  onChangeUniversities,
-}: {
-  appliedUniversities: AppliedUniversity[];
-  onChangeUniversities: () => void;
-}) => (
-  <section className="mt-5">
-    <div className="flex items-center justify-between gap-3">
-      <h1 className="text-k-900 typo-bold-4">지원한 대학 ({appliedUniversities.length}개)</h1>
-      <button
-        type="button"
-        onClick={onChangeUniversities}
-        className="h-8 shrink-0 rounded-2xl bg-primary px-4 text-k-0 typo-sb-12"
-      >
-        지원 대학교 변경
-      </button>
-    </div>
-    <div className="mt-4 space-y-2">
-      {appliedUniversities.length > 0 ? (
-        appliedUniversities.map(({ preference, scoreSheet }) => (
-          <AppliedUniversityRow
-            key={`${preference}-${scoreSheet.koreanName}`}
-            preference={preference}
-            scoreSheet={scoreSheet}
-          />
-        ))
-      ) : (
-        <div className="rounded-lg bg-k-50 px-4 py-4 text-k-500 typo-medium-3">
-          지원한 대학 정보를 불러오는 중입니다.
-        </div>
-      )}
-    </div>
-  </section>
-);
-
-const AppliedUniversityRow = ({ preference, scoreSheet }: AppliedUniversity) => {
-  const capacity =
-    scoreSheet.studentCapacity === null || scoreSheet.studentCapacity === undefined
-      ? "미정"
-      : scoreSheet.studentCapacity;
-
-  return (
-    <Link
-      href={getApplicationDetailHref(scoreSheet.koreanName)}
-      className="flex h-11 items-center gap-2.5 rounded-lg border border-k-50 bg-white px-3.5 shadow-[0_0_10px_rgba(0,0,0,0.05)]"
-    >
-      <ScoreSheetLogo scoreSheet={scoreSheet} className="size-5" />
-      <span className="min-w-0 flex-1 truncate text-k-900 typo-sb-7">
-        {scoreSheet.koreanName} ({preference}/{capacity})
-      </span>
-      <span className="flex size-6 shrink-0 items-center justify-center text-k-600" aria-hidden>
-        <IconExpandMoreFilled />
-      </span>
-    </Link>
-  );
-};
-
-const ParticipantBanner = ({ participantCount }: { participantCount: number }) => (
-  <section className="mt-6 rounded-lg bg-secondary-100 px-5 py-3">
-    <div className="flex items-center gap-4">
-      <div className="flex size-10 items-center justify-center text-[34px]" aria-hidden>
-        🔥
-      </div>
-      <div className="min-w-0 text-k-800">
-        <p className="typo-regular-4">솔리드 커넥션과 함께하고 있어요</p>
-        <p className="mt-0.5 typo-sb-9">총 {participantCount}명이 성적 공유 참여중!</p>
-      </div>
-    </div>
-  </section>
-);
-
-const ApplicationScopeTabs = ({
-  scope,
-  totalUniversityCount,
-  applicantUniversityCount,
-  onScopeChange,
-}: {
-  scope: ApplicantScope;
-  totalUniversityCount: number;
-  applicantUniversityCount: number;
-  onScopeChange: (scope: ApplicantScope) => void;
-}) => (
-  <div className="mt-6 flex h-11 items-start justify-between">
-    <ScopeTabButton isActive={scope === "all"} onClick={() => onScopeChange("all")}>
-      모든 대학 ({totalUniversityCount}개)
-    </ScopeTabButton>
-    <ScopeTabButton isActive={scope === "withApplicants"} onClick={() => onScopeChange("withApplicants")}>
-      지원자 있는 대학 ({applicantUniversityCount}개)
-    </ScopeTabButton>
-  </div>
-);
-
-const ScopeTabButton = ({
-  children,
-  isActive,
-  onClick,
-}: {
-  children: ReactNode;
-  isActive: boolean;
-  onClick: () => void;
-}) => (
-  <button
-    type="button"
-    onClick={onClick}
-    className={clsx(
-      "flex h-11 flex-1 items-center justify-center border-b-2 typo-medium-2",
-      isActive ? "border-primary text-k-900" : "border-transparent text-k-300",
-    )}
-  >
-    {children}
-  </button>
-);
-
-const ApplicationFilterChips = ({
-  regionFilter,
-  sortMode,
-  onRegionChange,
-  onSortModeChange,
-}: {
-  regionFilter: RegionKo | "";
-  sortMode: ScoreSort;
-  onRegionChange: (region: RegionKo | "") => void;
-  onSortModeChange: (sortMode: ScoreSort) => void;
-}) => (
-  <div className="mt-4 flex gap-2 overflow-x-auto whitespace-nowrap pb-1">
-    {REGIONS_KO.map((region) => (
-      <FilterChip
-        key={region}
-        isActive={regionFilter === region}
-        onClick={() => onRegionChange(regionFilter === region ? "" : region)}
-      >
-        {region}
-      </FilterChip>
-    ))}
-    <FilterChip
-      isActive={sortMode === "gpa"}
-      onClick={() => onSortModeChange(sortMode === "gpa" ? "applicants" : "gpa")}
-    >
-      학점 높은 순
-    </FilterChip>
-  </div>
-);
-
-const FilterChip = ({
-  children,
-  isActive,
-  onClick,
-}: {
-  children: ReactNode;
-  isActive: boolean;
-  onClick: () => void;
-}) => (
-  <button
-    type="button"
-    onClick={onClick}
-    className={clsx("rounded-full px-3 py-[5px] typo-sb-12", isActive ? "bg-primary text-k-0" : "bg-k-50 text-k-300")}
-  >
-    {children}
-  </button>
-);
-
-const ScoreSheetList = ({ scoreSheets }: { scoreSheets: ScoreSheetType[] }) => {
-  if (scoreSheets.length === 0) {
-    return <ApplicationScoreEmptyState />;
-  }
-
-  return (
-    <div className="mt-3 flex flex-col gap-2 pb-6">
-      {scoreSheets.map((scoreSheet, index) => (
-        <MobileScoreSheet key={scoreSheet.koreanName} scoreSheet={scoreSheet} defaultOpen={index === 0} />
-      ))}
-    </div>
-  );
-};
-
-const ApplicationScoreEmptyState = () => (
-  <div className="mt-4 flex min-h-48 flex-col items-center justify-center rounded-lg bg-k-50 px-6 py-8 text-center">
-    <p className="text-k-900 typo-sb-7">조건에 맞는 대학이 없어요.</p>
-    <p className="mt-2 text-k-500 typo-medium-3">다른 필터로 다시 확인해 주세요.</p>
-  </div>
-);
-
-const uniqueScoreSheets = (scoreSheets: ScoreSheetType[]) => {
-  const scoreSheetMap = new Map<string, ScoreSheetType>();
-
-  for (const scoreSheet of scoreSheets) {
-    const key = getScoreSheetKey(scoreSheet);
-    const prev = scoreSheetMap.get(key);
-    scoreSheetMap.set(
-      key,
-      prev
-        ? {
-            ...prev,
-            englishName: prev.englishName ?? scoreSheet.englishName,
-            logoImageUrl: prev.logoImageUrl ?? scoreSheet.logoImageUrl,
-            backgroundImageUrl: prev.backgroundImageUrl ?? scoreSheet.backgroundImageUrl,
-            applicants: mergeApplicants(prev.applicants, scoreSheet.applicants),
-          }
-        : scoreSheet,
-    );
-  }
-
-  return Array.from(scoreSheetMap.values());
-};
-
-const getAppliedUniversities = (scoreChoices: ScoreSheetType[][], maxChoiceCount: number): AppliedUniversity[] => {
-  return scoreChoices.slice(0, maxChoiceCount).flatMap((choice, index) => {
-    const mine = choice.find((scoreSheet) => scoreSheet.applicants.some((applicant) => applicant.isMine));
-
-    return mine ? [{ preference: index + 1, scoreSheet: mine }] : [];
+  const { data: gpaScoreData, isLoading: isGpaLoading } = useGetMyGpaScore({ enabled: shouldFetchScoreStatus });
+  const { data: languageTestScores = [], isLoading: isLanguageTestLoading } = useGetMyLanguageTestScore({
+    enabled: shouldFetchScoreStatus,
   });
-};
 
-const getScoreSheetKey = (scoreSheet: ScoreSheetType) => {
-  return scoreSheet.id ? String(scoreSheet.id) : scoreSheet.koreanName;
-};
-
-const mergeApplicants = (currentApplicants: Applicant[], nextApplicants: Applicant[]) => {
-  const applicantMap = new Map<string, Applicant>();
-
-  for (const applicant of currentApplicants) {
-    applicantMap.set(applicant.nicknameForApply, applicant);
+  if (!isAuthInitialized || (shouldFetchScoreStatus && (isGpaLoading || isLanguageTestLoading))) {
+    return <CloudSpinnerPage />;
   }
 
-  for (const applicant of nextApplicants) {
-    if (!applicantMap.has(applicant.nicknameForApply)) {
-      applicantMap.set(applicant.nicknameForApply, applicant);
-    }
-  }
-
-  return Array.from(applicantMap.values());
-};
-
-const getParticipantCount = (scoreSheets: ScoreSheetType[]) => {
-  const nicknames = new Set<string>();
-
-  for (const scoreSheet of scoreSheets) {
-    for (const applicant of scoreSheet.applicants) {
-      nicknames.add(applicant.nicknameForApply);
-    }
-  }
-
-  return nicknames.size;
-};
-
-const getAverageGpa = (scoreSheet: ScoreSheetType) => {
-  if (scoreSheet.applicants.length === 0) {
-    return 0;
-  }
-
-  const totalGpa = scoreSheet.applicants.reduce((sum, applicant) => sum + applicant.gpa, 0);
-  return totalGpa / scoreSheet.applicants.length;
-};
-
-const sortScoreSheets = (scoreSheets: ScoreSheetType[], sortMode: ScoreSort) => {
-  return [...scoreSheets].sort((a, b) => {
-    if (sortMode === "gpa") {
-      return getAverageGpa(b) - getAverageGpa(a);
-    }
-
-    return b.applicants.length - a.applicants.length;
+  const access = resolveApplicationStatusAccess({
+    isAuthenticated,
+    homeUniversityId,
+    gpaScores: gpaScoreData?.gpaScoreStatusResponseList ?? [],
+    languageTestScores,
   });
+
+  switch (access) {
+    case "guest":
+      return <GuestApplicationStatusPage />;
+    case "signedIn":
+      return <SignedInApplicationStatusPage />;
+    case "scorePending":
+      return <ScorePendingApplicationStatusPage />;
+    case "approved":
+      return <ApprovedApplicationStatusPage />;
+  }
 };
 
 export default ScorePageContent;
