+++ apps/web/src/app/university/application/_pages/SignedInApplicationStatusPage.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { useUniversitySearch } from "@/apis/universities";
+import RestrictedApplicationStatusView from "../_components/RestrictedApplicationStatusView";
+
+const SignedInApplicationStatusPage = () => {
+  const { data: universities = [], isLoading } = useUniversitySearch("", undefined, { useDefaultTermId: true });
+
+  return <RestrictedApplicationStatusView universities={universities} isLoading={isLoading} />;
+};
+
+export default SignedInApplicationStatusPage;
